+++ test/test_helper.rb
@@ -28,7 +28,6 @@ end
 
 $LOAD_PATH.unshift File.expand_path("../lib", __dir__)
 
-require "bundler/setup"
 Bundler.setup :default, :test
 
 require "minitest/autorun"